# swagger/railroad.yaml
openapi: 3.0.0
info:
  title: RailRoad API
  description: API for managing trains, stations, tickets, and users
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
paths:
...
  /trains:
    post:
      summary: Create a new train (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                start_station:
                  type: string
                end_station:
                  type: string
                time_of_departure:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Train ajouté
        '403':
          description: Accès refusé (seulement pour les admins)

    put:
      summary: Update a specific train (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                start_station:
                  type: string
                end_station:
                  type: string
                time_of_departure:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Train mis à jour
        '403':
          description: Accès refusé (seulement pour les admins)

    delete:
      summary: Delete a specific train (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Train supprimé
        '403':
          description: Accès refusé (seulement pour les admins)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT